Generic packet (#140)
* TMP: Temporary commit moving packet classes around

* ENH: Allow for binary data other than CCSDS Packets

This moves towards a generic packet interface and removes the need
for passing in bytes with a read_* interface, we now wrap that in
our own reader class to simplify that interaction.

* MNT: Use generic Packet interface in accepted types.

* MNT: Add header and user_data properties for backwards compatibility

We can add these properties back onto the main "Packet" class and
add a user warning when being accessed to make it easier to
transition to the new style and provide a helpful message.

* MNT: Add explicit parse_bytes and parse_packet methods to definition

Deprecate parse_ccsds_packet() as the XTCE spec doesn't need to know
about the ccsds standard. We should handle parsing raw bytes objects
and create the underlying objects for the users ourselves.

* MNT/TST: Add better error handling when reading beyond the end of data stream

* MNT: Update documentation/changelog for generic packet interface

* FIX: Line too long in tests

* MNT: Add ccsds apid to continuation packet warning message

* MNT: Change deprecations to use DeprecationWarning

* DOC: Update the ccsds_generator docstring

* MNT: pop packets once completed to clean up dictionary storage

* MNT: Major refactor of modules / locations / and classes

* DOC: Update documentation for RawPacketBytes and other methods that have changed

* MNT: Remove continuation packet from XTCE parsing

It is specific to ccsds packets and hasn't been released yet, so
just keep it over there instead.

* TST: Fixing up CCSDSPacket rename

* MNT: Rename Packet -> SpacePacket

* Add fixed length packet generator

* DOC: Change Packet to SpacePacket in types

* DOC: Update ccsds continuation packet description