FIX: Only yield complete packets from ccsds_generator (#218)
Previously, if we had enough bytes for a partial packet, we
could end up creating a packet that was not as long as the number
of bytes requested. This can happen if the file delivered has a
partial packet at the end of it. Addtionally, if we only have
a few bytes left at the end, we would try to read from the bytes
source and end up reading from None, so implement a pass-through
function to make sure that doesn't fail if it is called.